Move more tests to folder and fix tests
